+++ BattleBot/BattleBot/Components/SimpleTexture.cs
@@ -17,6 +17,6 @@ namespace BattleBot.Components
         public Texture2D Texture { get; internal set; }
         public Color Tint { get; internal set; }
 
-        public int Depth {  get; internal set; }
+        public float Depth {  get; internal set; }
     }
 }
